pyproject.toml
[project]
    authors = [{ name = "Nate Landau", email = "github@natelandau.com" }]
    dependencies = [
        "aerich[toml]>=0.8.1",
        "aiohttp>=3.13.3",
        "aiosqlite==0.21.0",
        "audioop-lts>=0.2.2,<1",
        "httpx>=0.28.1",
        "humanize>=4.15.0",
        "inflect>=7.5.0",
        "loguru>=0.7.3",
        "py-cord>=2.7.1",
        "pydantic-settings>=2.13.1",
        "pydantic[email]>=2.12.5",
        "rich>=14.3.3",
        "semver>=3.0.4",
        "tortoise-orm>=1.1.5",
        "valentina-python-client>=1.5.0",
    ]
    description = "A Discord bot for the Valentina Noir ecosystem."
    license = { file = "LICENSE" }
    name = "vbot"
    readme = "README.md"
    requires-python = ">=3.13"
    version = "0.1.0"

    [project.scripts]
        vbot = "vbot.run:main"

[build-system]
    build-backend = "uv_build"
    requires      = ["uv_build>=0.9.4,<0.10.0"]

[dependency-groups]
    dev = [
        "anyio>=4.12.1",
        "commitizen>=4.13.8",
        "duty>=1.9.0",
        "mypy>=1.19.1",
        "prek>=0.3.3",
        "pytest-clarity>=1.0.1",
        "pytest-cov>=7.0.0",
        "pytest-devtools>=1.0.0",
        "pytest-mock>=3.15.1",
        "pytest-repeat>=0.9.4",
        "pytest-sugar>=1.1.1",
        "pytest-xdist>=3.8.0",
        "ruff>=0.15.2",
        "shellcheck-py>=0.11.0.1",
        "typos>=1.43.5",
        "yamllint>=1.38.0",
    ]

[tool.commitizen]
    bump_message               = "bump(release): v$current_version → v$new_version"
    changelog_merge_prerelease = true
    tag_format                 = "v$version"
    update_changelog_on_bump   = true
    version_files              = ["src/vbot/__init__.py:__version__"]
    version_provider           = "uv"

[tool.coverage.report] # https://coverage.readthedocs.io/en/latest/config.html#report
    exclude_lines = [
        'def __repr__',
        'except [\w\s\._]+ as .*:',
        'if TYPE_CHECKING',
        'pragma: no cover',
        'raise Abort',
        'raise Exit',
    ]
    fail_under = 0
    precision = 1
    show_missing = true
    skip_covered = true
    skip_empty = true

[tool.coverage.run]
    branch       = true
    command_line = "--module pytest"
    data_file    = ".cache/coverage"
    omit         = ["tests/*"]
    source       = ["src"]

[tool.coverage.xml]
    output = ".cache/coverage.xml"

[tool.mypy] # https://mypy.readthedocs.io/en/latest/config_file.html
    cache_dir                   = ".cache/mypy"
    disallow_any_unimported     = false
    disallow_subclassing_any    = false
    disallow_untyped_decorators = false
    disallow_untyped_defs       = true
    exclude                     = ['duties.py', 'tests/']
    follow_imports              = "normal"
    ignore_missing_imports      = true
    junit_xml                   = ".cache/mypy.xml"
    no_implicit_optional        = true
    pretty                      = false
    show_column_numbers         = true
    show_error_codes            = true
    show_error_context          = true
    strict_optional             = false
    warn_redundant_casts        = true
    warn_unreachable            = true
    warn_unused_ignores         = true

[tool.pytest.ini_options]

    addopts = "--color=yes --doctest-modules --exitfirst --failed-first --strict-config --strict-markers"
    cache_dir = ".cache/pytest"
    filterwarnings = [
        'error',
        'ignore:.*Pydantic.*:UserWarning',
        'ignore::DeprecationWarning:',
        'ignore::pytest.PytestUnraisableExceptionWarning',
        'ignore:unclosed.*:ResourceWarning',
    ]
    markers = [
        "drop_db: drops the database before the test creating fresh data",
        "no_db: marks tests as not needing a database",
        "serial",
    ]
    testpaths = ["tests"]
    xfail_strict = true

[tool.ruff] # https://github.com/charliermarsh/ruff

    exclude        = [".cache", ".git", ".venv", "_build", "build", "dist", "src/vbot/db/migrations"]
    fix            = true
    line-length    = 100
    output-format  = "grouped"
    src            = ["src", "tests"]
    target-version = "py313"
    [tool.ruff.lint]
        ignore = [
            "ANN002",  # Missing type annotation for `*args`
            "ANN003",  # Missing type annotation for `**kwargs`
            "ANN204",  # missing return type annotation for special method `__init__`
            "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed,
            "B006",    # mutable-argument-default
            "B008",    # function-call-in-default-argument
            "COM812",  # Trailing comma missing"
            "CPY001",  # Missing copyright notice at top of file
            "D107",    # undocumented-public-init
            "E501",    # line-too-long
            "PLC0415", # Imports should be at top of file
            # "FBT001", # Boolean-typed positional argument in function definition
            # "FBT002", # Boolean-typed positional argument in function definition
            "FIX002", # Line contains TODO, consider resolving the issue
            "S311",   # suspicious-non-cryptographic-random-usage
            "TD001",  # invalid-todo-tag
            "TD002",  # Missing author in TODO
            "TD003",  # Missing issue link on the line following this TODO
        ]
        per-file-ignores = { "tests/**/*.py" = [
            "A002",
            "A003",
            "ANN001",  # Missing type annotation for function argument `cls`
            "ANN002",  # Missing type annotation for `*args`
            "ANN003",  # Missing type annotation for `**kwargs`
            "ANN201",  # Missing return type annotation
            "ARG001",  # Unused function argument
            "ARG002",  # Unused method argument
            "ARG005",  # Unused lambda argument
            "D102",
            "ERA001",  # Commented out code
            "F403",
            "F405",    # May be undefined from type imports
            "PGH003",  # Use specific rule codes when ignoring type issues
            "PLC0415", # Imports should be at top of file
            "PLR0913",
            "PLR2004",
            "S101",
            "SLF001",  # Calling private method
            "W292",    # No blank line at end of file - included b/c cursor struggles to add a trailing newline and burns through requests trying to fix this linting error.
        ] }
        select = ["ALL"]
        unfixable = [
            "ERA001", # Commented out code
            "F401",   # unused-import
            "F841",   # unused-variable
        ]

        [tool.ruff.lint.mccabe]
            # Unlike Flake8, default to a complexity level of 10.
            max-complexity = 10

        [tool.ruff.lint.pydocstyle]
            convention = "google"

        [tool.ruff.lint.pylint]
            max-args = 6

    [tool.ruff.format]
        indent-style              = "space"
        line-ending               = "auto"
        quote-style               = "double"
        skip-magic-trailing-comma = false
